Fixed problem caused by 1.0 schema change.
This fix is because of a change to the 1.0 schema which broke a query
in annotation.R.  The 1.0 schema changes the internal id col of all
the annotation packages from "id" to "_id".  This package depends on
these schemas for an example, which is why this package broke and why
I poked my nose into Colins package.  It's because I felt responsible
for the breakage.  Since I put the 1.0 schemas into place recently.
Hopefully the schemas will not change again for a very very long time.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/annaffy@29460 bc3139a8-67e5-0310-9ffc-ced21a209358